requirements.txt
asgiref==3.5.2
Django==3.2.7
django-braces==1.14.0
django-crispy-forms==1.9.1
fontawesomefree==6.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.4
pandas==1.4.2
protobuf==4.21.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
xlrd==2.0.1
wfastcgi==3.0.0